+++ fresfolio/static/js/vue_assets/compositions/project_layout.js
@@ -491,8 +491,8 @@ const ProjectLayout = defineComponent({
             if (sectionPinned) {
                 this.pinnedSections = this.pinnedSections.filter(section => !(section.ID === sectionID && section.projectID === projectID));
             } else {
-                sectionInRenderedSections = this.renderedSections.find(section => section.ID === sectionID && section.projectID === projectID);
-                sectionInSearchSections = this.searchSections.find(section => section.ID === sectionID && section.projectID === projectID);
+                const sectionInRenderedSections = this.renderedSections.find(section => section.ID === sectionID && section.projectID === projectID);
+                const sectionInSearchSections = this.searchSections.find(section => section.ID === sectionID && section.projectID === projectID);
                 if (sectionInRenderedSections) {
                     const sectionJSON = this.renderedSections.find(section => section.ID === sectionID && section.projectID === projectID);
                     this.pinnedSections.push({ ...sectionJSON })