+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 This is a bug fix release for various issues discovered after the release v0.3.2. Also new features have been added.
 
 ## What's Changed
-* FEATURE: Omilayers can be plotted with vgplot.
+* FEATURE: Omilayers can be plotted with vgplot.js.
+* FEATURE: Omilayers tables can be viewed with perspective.js.